+++ cone.h
@@ -16,11 +16,11 @@ class cone: public shape
 	{
 		center=c;
 		shapeTransform.setIdentity();
-		minimum=-1*std::numeric_limits<double>::infinity();
-		maximum=std::numeric_limits<double>::infinity();
+		minimum=-1*std::numeric_limits<float>::infinity();
+		maximum=std::numeric_limits<float>::infinity();
 		closed=cl;
 	}
-	cone(const tuple& c, const double& min, const double& max, const bool& cl)
+	cone(const tuple& c, const float& min, const float& max, const bool& cl)
 	{
 		center=c;
 		shapeTransform.setIdentity();
@@ -28,42 +28,42 @@ class cone: public shape
 		maximum=max;
 		closed=cl;
 	}
-	double ray_hits_me(const ray& r, double& near_hit_point){
+	float ray_hits_me(const ray& r, float& near_hit_point){
 		matrix inv=invertMatrix(this->shapeTransform);
 		tuple transformRayOrigin=r.origin()*inv;
 		tuple transformRayDestination=r.direction()*inv;
-		double a= pow(transformRayDestination.x(),2)-pow(transformRayDestination.y(),2)+pow(transformRayDestination.z(),2);
-		double b= 2*transformRayOrigin.x()*transformRayDestination.x()+
+		float a= pow(transformRayDestination.x(),2)-pow(transformRayDestination.y(),2)+pow(transformRayDestination.z(),2);
+		float b= 2*transformRayOrigin.x()*transformRayDestination.x()+
 				  2*transformRayOrigin.z()*transformRayDestination.z()-
 				  2*transformRayOrigin.y()*transformRayDestination.y();
-		double c= pow(transformRayOrigin.x(),2)+pow(transformRayOrigin.z(),2)-pow(transformRayOrigin.y(),2);
+		float c= pow(transformRayOrigin.x(),2)+pow(transformRayOrigin.z(),2)-pow(transformRayOrigin.y(),2);
 		if (abs(a)<=0){ 
 			if (abs(b)<=0){
 				return intersect_caps(transformRayOrigin,transformRayDestination,near_hit_point);
 			}	
 			else{
-				double t= -c/(2*b);	
-				double y0= transformRayOrigin.y()+t*transformRayDestination.y();
+				float t= -c/(2*b);	
+				float y0= transformRayOrigin.y()+t*transformRayDestination.y();
 				if(minimum<y0 && y0<maximum && t>=0 && t<near_hit_point){
 					near_hit_point=t;
 				}		
 			}	
 		}
 		else{
-			double disc= pow(b,2)-4*a*c;
+			float disc= pow(b,2)-4*a*c;
 			if (disc<0){
 				return near_hit_point;
 			}
-			double t0= (-b-sqrt(disc))/(2*a);
-			double t1= (-b+sqrt(disc))/(2*a);
+			float t0= (-b-sqrt(disc))/(2*a);
+			float t1= (-b+sqrt(disc))/(2*a);
 			if(t0>t1){
 				std::swap(t0,t1);
 			}
-			double y0= transformRayOrigin.y()+t0*transformRayDestination.y();
+			float y0= transformRayOrigin.y()+t0*transformRayDestination.y();
 			if(minimum<y0 && y0<maximum && t0>=0 && t0<near_hit_point){
 				near_hit_point=t0;
 			}
-			double y1= transformRayOrigin.y()+t1*transformRayDestination.y();
+			float y1= transformRayOrigin.y()+t1*transformRayDestination.y();
 			if(minimum<y1 && y1<maximum && t1>=0 && t1<near_hit_point){
 				near_hit_point=t1;
 			}
@@ -74,7 +74,7 @@ class cone: public shape
 	tuple normal_at(tuple& point){
 		matrix inv = invertMatrix(this->shapeTransform);
 		tuple object_point=point*inv;
-		double dist= (pow(object_point.x(),2)+pow(object_point.z(),2));
+		float dist= (pow(object_point.x(),2)+pow(object_point.z(),2));
 		tuple object_normal;
 		if(dist<1 and object_point.y()>=maximum-EPSILON2){
 			object_normal=tuple(0,1,0,0);		}
@@ -82,7 +82,7 @@ class cone: public shape
 			object_normal=tuple(0,-1,0,0);	
 		}
 		else{
-			double y = sqrt(pow(object_point.x(),2)+pow(object_point.z(),2));
+			float y = sqrt(pow(object_point.x(),2)+pow(object_point.z(),2));
 			if (object_point.y()>0){
 				y=-y;
 			}
@@ -95,28 +95,28 @@ class cone: public shape
 		//std::cout << u.x() << " " << u.y() << " " << u.z() << " world_normal \n";
 		return u;
 	};
-	bool check_cap(const tuple& origin, const tuple& direction, double& t, double& y){
-		double x = origin.x()+t*direction.x();
-		double z = origin.z()+t*direction.z();
+	bool check_cap(const tuple& origin, const tuple& direction, float& t, float& y){
+		float x = origin.x()+t*direction.x();
+		float z = origin.z()+t*direction.z();
 		return (pow(x,2)+pow(z,2))<= pow(y,2);
 	}
-	double intersect_caps(const tuple& origin, const tuple& direction,double& near_hit_point){
+	float intersect_caps(const tuple& origin, const tuple& direction,float& near_hit_point){
 		if(closed==false || abs(direction.y())<EPSILON2){
 			return near_hit_point;
 		}
-		double t0=(minimum-origin.y())/direction.y();
+		float t0=(minimum-origin.y())/direction.y();
 		//std::cout << t0<<" t3 \n";
 		if(check_cap(origin,direction,t0,minimum) && t0>=0 && t0<near_hit_point){
 			near_hit_point=t0;
 		}
-		double t1=(maximum-origin.y())/direction.y();
+		float t1=(maximum-origin.y())/direction.y();
 		if(check_cap(origin,direction,t1,maximum) && t1>=0 && t1<near_hit_point){
 			near_hit_point=t1;
 		}
 		return near_hit_point;		
 	}
-	double minimum;
-	double maximum;
+	float minimum;
+	float maximum;
 	bool closed;
 };
